Python-based Text Analyzer
The project requires the creation of two programs: single thread program and multi-thread program.  Both programs should output the following: number of words, number of sentences, and number of characters (excluding spaces).  The input for both programs is a text file.   Sample input text files can be taken from: https://www.gutenberg.org/

For the multi-threaded program, it should use multiple threads in finding the statistics.  There should be at least two threads being used for determining the number of words and sentences.  The program should show the number of threads in use and how the threads are is accessing the file and semaphores or locks in determining the statistics.  

The program created can be tested in comparison with an online text analyzer at https://www.online-utility.org/text (Links to an external site.)analyzer.jsp